test/regression/reftest/web-platform-tests/dom/events/Event-type.html

<!DOCTYPE html>
<title>Event.type</title>
<link rel="author" title="Ms2ger" href="mailto:Ms2ger@gmail.com">
<link rel="help" href="https://dom.spec.whatwg.org/#dom-event-type">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  // createEvent is not supported
  // var e = document.createEvent("Event")
  var e = new Event("")
  assert_equals(e.type, "");
}, "Event.type should initially be the empty string");
test(function() {
  // createEvent and initEvent are not supported
  // var e = document.createEvent("Event")
  // e.initEvent("foo", false, false)
  var e = new Event("foo", {"bubbles":false, "cancelable":false})
  assert_equals(e.type, "foo")
}, "Event.type should be initialized by initEvent");
test(function() {
  var e = new Event("bar")
  assert_equals(e.type, "bar")
}, "Event.type should be initialized by the constructor");
</script>
